Closes #7 - activate authentication Improve and fix tests
