+++ test/test_module_event.py
@@ -1,0 +1,334 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2015-2016: Alignak team, see AUTHORS.txt file for contributors
+#
+# This file is part of Alignak.
+#
+# Alignak is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Alignak is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Alignak.  If not, see <http://www.gnu.org/licenses/>.
+#
+"""
+Test the module
+"""
+
+import os
+import re
+import time
+import json
+
+import shlex
+import subprocess
+
+import logging
+
+import requests
+
+from alignak_test import AlignakTest, time_hacker
+from alignak.modulesmanager import ModulesManager
+from alignak.objects.module import Module
+from alignak.basemodule import BaseModule
+
+# Set environment variable to ask code Coverage collection
+os.environ['COVERAGE_PROCESS_START'] = '.coveragerc'
+
+import alignak_module_ws
+
+# # Activate debug logs for the alignak backend client library
+# logging.getLogger("alignak_backend_client.client").setLevel(logging.DEBUG)
+#
+# # Activate debug logs for the module
+# logging.getLogger("alignak.module.web-services").setLevel(logging.DEBUG)
+
+
+class TestModuleWs(AlignakTest):
+    """This class contains the tests for the module"""
+
+    @classmethod
+    def setUpClass(cls):
+
+        # Set test mode for alignak backend
+        os.environ['TEST_ALIGNAK_BACKEND'] = '1'
+        os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-module-ws-backend-test'
+
+        # Delete used mongo DBs
+        print ("Deleting Alignak backend DB...")
+        exit_code = subprocess.call(
+            shlex.split(
+                'mongo %s --eval "db.dropDatabase()"' % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
+        )
+        assert exit_code == 0
+
+        fnull = open(os.devnull, 'w')
+        cls.p = subprocess.Popen(['uwsgi', '--plugin', 'python', '-w', 'alignakbackend:app',
+                                  '--socket', '0.0.0.0:5000',
+                                  '--protocol=http', '--enable-threads', '--pidfile',
+                                  '/tmp/uwsgi.pid'],
+                                 stdout=fnull, stderr=fnull)
+        time.sleep(3)
+
+        endpoint = 'http://127.0.0.1:5000'
+
+        test_dir = os.path.dirname(os.path.realpath(__file__))
+        print("Current test directory: %s" % test_dir)
+
+        print("Feeding Alignak backend... %s" % test_dir)
+        exit_code = subprocess.call(
+            shlex.split('alignak-backend-import --delete %s/cfg/cfg_default.cfg' % test_dir),
+            stdout=fnull, stderr=fnull
+        )
+        assert exit_code == 0
+        print("Fed")
+
+        # Backend authentication
+        headers = {'Content-Type': 'application/json'}
+        params = {'username': 'admin', 'password': 'admin'}
+        # Get admin user token (force regenerate)
+        response = requests.post(endpoint + '/login', json=params, headers=headers)
+        resp = response.json()
+        cls.token = resp['token']
+        cls.auth = requests.auth.HTTPBasicAuth(cls.token, '')
+
+        # Get admin user
+        response = requests.get(endpoint + '/user', auth=cls.auth)
+        resp = response.json()
+        cls.user_admin = resp['_items'][0]
+
+        # Get realms
+        response = requests.get(endpoint + '/realm', auth=cls.auth)
+        resp = response.json()
+        cls.realmAll_id = resp['_items'][0]['_id']
+
+        # Add a user
+        data = {'name': 'test', 'password': 'test', 'back_role_super_admin': False,
+                'host_notification_period': cls.user_admin['host_notification_period'],
+                'service_notification_period': cls.user_admin['service_notification_period'],
+                '_realm': cls.realmAll_id}
+        response = requests.post(endpoint + '/user', json=data, headers=headers,
+                                 auth=cls.auth)
+        resp = response.json()
+        print("Created a new user: %s" % resp)
+
+    @classmethod
+    def tearDownClass(cls):
+        cls.p.kill()
+
+    def test_module_zzz_event(self):
+        """Test the module /event endpoint
+        :return:
+        """
+        self.print_header()
+        # Obliged to call to get a self.logger...
+        self.setup_with_file('cfg/cfg_default.cfg')
+        self.assertTrue(self.conf_is_correct)
+
+        # -----
+        # Provide parameters - logger configuration file (exists)
+        # -----
+        # Clear logs
+        self.clear_logs()
+
+        # Create an Alignak module
+        mod = Module({
+            'module_alias': 'web-services',
+            'module_types': 'web-services',
+            'python_name': 'alignak_module_ws',
+            # Alignak backend
+            'alignak_backend': 'http://127.0.0.1:5000',
+            'username': 'admin',
+            'password': 'admin',
+            # Set Arbiter address as empty to not poll the Arbiter else the test will fail!
+            'alignak_host': '',
+            'alignak_port': 7770,
+        })
+
+        # Create the modules manager for a daemon type
+        self.modulemanager = ModulesManager('receiver', None)
+
+        # Load an initialize the modules:
+        #  - load python module
+        #  - get module properties and instances
+        self.modulemanager.load_and_init([mod])
+
+        my_module = self.modulemanager.instances[0]
+
+        # Clear logs
+        self.clear_logs()
+
+        # Start external modules
+        self.modulemanager.start_external_instances()
+
+        # Starting external module logs
+        self.assert_log_match("Trying to initialize module: web-services", 0)
+        self.assert_log_match("Starting external module web-services", 1)
+        self.assert_log_match("Starting external process for module web-services", 2)
+        self.assert_log_match("web-services is now started", 3)
+
+        # Check alive
+        self.assertIsNotNone(my_module.process)
+        self.assertTrue(my_module.process.is_alive())
+
+        time.sleep(1)
+
+        # ---
+        # Prepare the backend content...
+        self.endpoint = 'http://127.0.0.1:5000'
+
+        headers = {'Content-Type': 'application/json'}
+        params = {'username': 'admin', 'password': 'admin'}
+        # get token
+        response = requests.post(self.endpoint + '/login', json=params, headers=headers)
+        resp = response.json()
+        self.token = resp['token']
+        self.auth = requests.auth.HTTPBasicAuth(self.token, '')
+
+        # Get default realm
+        response = requests.get(self.endpoint + '/realm', auth=self.auth)
+        resp = response.json()
+        self.realm_all = resp['_items'][0]['_id']
+        # ---
+
+        # Do not allow GET request on /event - not yet authorized
+        response = requests.get('http://127.0.0.1:8888/event')
+        self.assertEqual(response.status_code, 401)
+
+        session = requests.Session()
+
+        # Login with username/password (real backend login)
+        headers = {'Content-Type': 'application/json'}
+        params = {'username': 'admin', 'password': 'admin'}
+        response = session.post('http://127.0.0.1:8888/login', json=params, headers=headers)
+        assert response.status_code == 200
+        resp = response.json()
+
+        # Do not allow GET request on /event
+        response = session.get('http://127.0.0.1:8888/event')
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result['_status'], 'ERR')
+        self.assertEqual(result['_issues'], ['You must only POST on this endpoint.'])
+
+        self.assertEqual(my_module.received_commands, 0)
+
+        # You must have parameters when POSTing on /event
+        headers = {'Content-Type': 'application/json'}
+        data = {}
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result['_status'], 'ERR')
+        self.assertEqual(result['_issues'], ['You must POST parameters on this endpoint.'])
+
+        self.assertEqual(my_module.received_commands, 0)
+
+        # Notify an host event - missing host or service
+        headers = {'Content-Type': 'application/json'}
+        data = {
+            "fake": ""
+        }
+        self.assertEqual(my_module.received_commands, 0)
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result, {'_status': 'ERR', '_issues': ['Missing host and/or service parameter.']})
+
+        # Notify an host event - missing comment
+        headers = {'Content-Type': 'application/json'}
+        data = {
+            "host": "test_host",
+        }
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result, {'_status': 'ERR',
+                                  '_issues': ['Missing comment. If you do not have any comment, '
+                                              'do not comment ;)']})
+
+        # Notify an host event - default author
+        headers = {'Content-Type': 'application/json'}
+        data = {
+            "host": "test_host",
+            "comment": "My comment"
+        }
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result, {'_status': 'OK',
+                                  '_result': [u'ADD_HOST_COMMENT;test_host;1;'
+                                              u'Alignak WS;My comment']})
+
+        # Notify an host event - default author and timestamp
+        headers = {'Content-Type': 'application/json'}
+        data = {
+            "timestamp": 1234567890,
+            "host": "test_host",
+            "author": "Me",
+            "comment": "My comment"
+        }
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result, {'_status': 'OK',
+                                  '_result': [u'[1234567890] ADD_HOST_COMMENT;test_host;1;'
+                                              u'Me;My comment']})
+
+        # Notify a service event - default author
+        headers = {'Content-Type': 'application/json'}
+        data = {
+            "host": "test_host",
+            "service": "test_service",
+            "comment": "My comment"
+        }
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result, {'_status': 'OK',
+                                  '_result': [u'ADD_SVC_COMMENT;test_host;test_service;1;'
+                                              u'Alignak WS;My comment']})
+
+        # Notify a service event - default author and timestamp
+        headers = {'Content-Type': 'application/json'}
+        data = {
+            "timestamp": 1234567890,
+            "host": "test_host",
+            "service": "test_service",
+            "author": "Me",
+            "comment": "My comment"
+        }
+        response = session.post('http://127.0.0.1:8888/event', json=data, headers=headers)
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result, {'_status': 'OK',
+                                  '_result': [u'[1234567890] ADD_SVC_COMMENT;test_host;test_service;'
+                                              u'1;Me;My comment']})
+
+        # Get history to confirm that backend is ready
+        # ---
+        response = session.get(self.endpoint + '/history', auth=self.auth,
+                                params={"sort": "-_id", "max_results": 25, "page": 1})
+        resp = response.json()
+        print("Response: %s" % resp)
+        for item in resp['_items']:
+            assert item['type'] in ['webui.comment']
+        # Got 4 notified events, so we get 4 comments in the backend
+        self.assertEqual(len(resp['_items']), 4)
+        # ---
+
+        # Logout
+        response = session.get('http://127.0.0.1:8888/logout')
+        self.assertEqual(response.status_code, 200)
+        result = response.json()
+        self.assertEqual(result['_status'], 'OK')
+        self.assertEqual(result['_result'], 'Logged out')
+
+        self.modulemanager.stop_all()
